changing to using a floating map
I thought map would return a float. It doesn't. 

For the read in, using a custom implementation of map that does (mapf)
For the pulse out, shifting the scaling to 1k range since that's all the precision we need (multiplying by a thousand on input and allowing the rest to get truncated).